refactor: Removed explicit UI for pipeline sections and simply display the raw file instead
