+++ crates/bld_ui/src/app.rs
@@ -26,8 +26,8 @@ pub fn App() -> impl IntoView {
             <div class="h-full grid place-items-center">{move || app_dialog_content.get()}</div>
         </dialog>
         <Router>
-            <div class="h-screen bg-slate-900">
-                <div class="h-screen flex bg-grid">
+            <div class="h-screen bg-slate-800">
+                <div class="h-screen flex">
                     <Routes>
                         <Route path="/" view=Home>
                             <Route path="/" view=Dashboard/>
