+++ crates/bld_ui/src/components/card.rs
@@ -2,6 +2,6 @@ use leptos::*;
 
 #[component]
 pub fn Card(#[prop(into, optional)] class: String, children: Children) -> impl IntoView {
-    let class = format!("bg-slate-700 rounded-xl flex flex-col gap-4 {class}");
+    let class = format!("bg-slate-700 rounded-xl flex flex-col gap-4 border border-slate-600 {class}");
     view! { <div class=class>{children()}</div> }
 }
