+++ crates/bld_ui/src/pages/home/pipelines/info/artifacts.rs
