--- crates/bld_ui/src/pages/home/pipelines/v2/jobs.rs
@@ -1,123 +1,0 @@
-use crate::components::{
-    badge::Badge,
-    card::Card,
-    list::List,
-    tabs::{Tab, Tabs},
-};
-use bld_runner::step::v2::BuildStep;
-use leptos::{leptos_dom::logging, *};
-use std::collections::HashMap;
-
-#[component]
-pub fn PipelineJobsV2(
-    #[prop(into)] jobs: Signal<HashMap<String, Vec<BuildStep>>>,
-) -> impl IntoView {
-    let items_cache = create_rw_signal(HashMap::new());
-    let selected_tab = create_rw_signal(String::new());
-
-    let jobs = move || {
-        let data = jobs
-            .get()
-            .into_iter()
-            .map(|(k, v)| {
-                (
-                    k,
-                    v.into_iter()
-                        .map(|x| {
-                            serde_yaml::to_string(&x)
-                                .map_err(|e| logging::console_error(&format!("{:?}", e)))
-                                .unwrap_or_default()
-                        })
-                        .collect::<Vec<String>>(),
-                )
-            })
-            .collect::<HashMap<String, Vec<String>>>();
-
-        data
-    };
-
-    let items = move || {
-        let key = selected_tab.get();
-
-        if !items_cache.get_untracked().contains_key(&key) {
-            let key = key.clone();
-
-            let steps: Vec<(String, String)> = jobs()
-                .get(&key)
-                .cloned()
-                .unwrap_or_default()
-                .into_iter()
-                .enumerate()
-                .map(|(i, x)| (format!("{key}_{i}"), x))
-                .collect();
-
-            items_cache.update_untracked(|x| {
-                x.insert(key, steps);
-            });
-        }
-
-        items_cache
-            .get_untracked()
-            .get(&key)
-            .cloned()
-            .unwrap_or_default()
-    };
-
-    let _ = watch(
-        move || jobs(),
-        move |x, _, _| {
-            selected_tab.set(x.keys().next().cloned().unwrap_or_default());
-        },
-        true,
-    );
-
-    view! {
-        <Card class="min-h-full">
-            <div class="flex flex-col px-8 py-12 gap-y-4">
-                <div class="flex flex-col">
-                    <div class="text-xl">"Jobs"</div>
-                    <div class="text-gray-400">"The parallel jobs for this pipeline."</div>
-                </div>
-                <Show
-                    when=move || !jobs().is_empty()
-                    fallback=|| {
-                        view! {
-                            <div class="grid justify-items-center">
-                                <Badge>"No jobs configured."</Badge>
-                            </div>
-                        }
-                    }
-                >
-
-                    <Tabs>
-                        <For
-                            each=move || {
-                                jobs()
-                                    .into_keys()
-                                    .enumerate()
-                                    .map(|(i, x)| (i, x.clone(), x.clone(), x))
-                            }
-
-                            key=|(i, _, _, _)| *i
-                            let:child
-                        >
-                            <Tab
-                                is_selected=move || selected_tab.get() == *child.1
-                                on:click=move |_| selected_tab.set(child.2.to_owned())
-                            >
-                                {child.3}
-                            </Tab>
-                        </For>
-                    </Tabs>
-                    <List>
-                        <For each=move || items() key=|(k, _)| k.clone() let:child>
-                            <pre class="text-sm text-gray-200 p-4 rounded-lg bg-slate-800">
-                                {child.1}
-                            </pre>
-                        </For>
-                    </List>
-                </Show>
-            </div>
-        </Card>
-    }
-}
